refactor(schema): schedule dir
