+++ apps/backend/src/db/schema/index.ts
@@ -1,4 +1,4 @@
 export * from './auth'
-export * from './course'
-export * from './schedule'
+export * from './schedule/course'
+export * from './schedule/schedule'
 export * from './subscription'
