+++ apps/backend/src/db/schema/schedule/index.ts
@@ -1,0 +1,2 @@
+export * from './schedule';
+export * from './course';
